feat(mvp): full platform launch — all 18 services compile, hosting infra complete
Backend:
- Fix all Rust compile errors across 18 microservices (sqlx macros → runtime
  form, rdkafka BorrowedHeaders, JwtService API, AppError struct variants,
  require_permission! macro, TenantId/Uuid type mismatches, missing stub modules)
- payments/src/main.rs: fix crate reference (logisticos_payments::bootstrap)
- compliance: convert all 4 db repos from sqlx::query! to runtime .bind() form;
  fix admin_routes.rs require_permission method → macro invocation
- All 18 services pass cargo check --workspace clean

Hosting infra:
- Add Dockerfiles for all 4 Next.js portals (Node 20 Alpine, 3-stage, non-root)
- Add output: standalone to all 4 next.config files (required for Docker builds)
- Add Helm chart for compliance service (port 8017, Vault-injected secrets)
- Rewrite ci-frontend.yml: pnpm → npm, add standalone build verification
- Rewrite deploy-staging.yml: matrix strategy covering all 18 services with
  GHCR image build + Helm deploy + smoke tests
- Add deny.toml for cargo-deny license/ban policy (blocks GPL, bans native-tls)
- Add terraform modules: clickhouse (Altinity operator on EKS) and timescaledb
  (RDS PostgreSQL + custom parameter group); wire both into dev environment
- Fix .env port mismatches to match docker-compose (payments=8012, fleet=8008,
  hub-ops=8009, ai-layer=8016)
- Add compliance and ai_layer schemas to scripts/db/init.sql

+++ services/dispatch/src/infrastructure/db/compliance_cache.rs
@@ -2,7 +2,7 @@ use redis::AsyncCommands;
 use uuid::Uuid;
 
 const KEY_PREFIX: &str = "compliance:status:";
-const TTL_SECONDS: usize = 300; // 5 min — refreshed on every compliance.status_changed event
+const TTL_SECONDS: u64 = 300; // 5 min — refreshed on every compliance.status_changed event
 
 pub struct ComplianceCache {
     redis: redis::aio::ConnectionManager,
